+++ src/lsp/workspace.rs
@@ -410,8 +410,8 @@ impl Workspace {
             }
             for keyword in [
                 "assert", "await", "branch", "break", "continue", "copy", "false", "func",
-                "import", "let", "loop", "move", "pub", "ref", "remote", "return", "true", "type",
-                "enum", "try",
+                "import", "let", "loop", "move", "not", "pub", "ref", "remote", "return", "true",
+                "type", "enum", "try",
             ] {
                 insert_completion(&mut items, keyword, CompletionItemKind::KEYWORD, None);
             }